+++ src/components/app/app.tsx
@@ -1,20 +1,20 @@
-import { useEffect } from "react";
-import { useDispatch } from "react-redux";
-import { Route, Routes, useLocation } from "react-router-dom";
-import Layout from "../layout/layout";
-import BurgerPage from "../../pages/burger-page/burger-page";
-import LoginPage from "../../pages/login-page/login-page";
-import RegisterPage from "../../pages/register-page/register-page";
-import ForgetPasswordPage from "../../pages/forgot-password-page/forgot-password-page";
-import ResetPasswordPage from "../../pages/reset-password-page/reset-password-page";
-import ProfilePage from "../../pages/profile-page/profile-page";
-import NotFoundPage from "../../pages/not-found-page/not-found-page";
-import ProtectedRoute from "../protected-route/protected-route";
-import IngredientDetailsPage from "../../pages/ingredient-details-page/ingredient-details-page";
-import ProfileUserForm from "../profile-user-form/profile-user-form";
-import ProfileOrders from "../profile-order/profile-order";
-import IngredientDetails from "../ingredient-details/ingredient-details";
-import Modal from "../modal/modal";
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { Route, Routes, useLocation } from 'react-router-dom';
+import Layout from '../layout/layout';
+import BurgerPage from '../../pages/burger-page/burger-page';
+import LoginPage from '../../pages/login-page/login-page';
+import RegisterPage from '../../pages/register-page/register-page';
+import ForgetPasswordPage from '../../pages/forgot-password-page/forgot-password-page';
+import ResetPasswordPage from '../../pages/reset-password-page/reset-password-page';
+import ProfilePage from '../../pages/profile-page/profile-page';
+import NotFoundPage from '../../pages/not-found-page/not-found-page';
+import ProtectedRoute from '../protected-route/protected-route';
+import IngredientDetailsPage from '../../pages/ingredient-details-page/ingredient-details-page';
+import ProfileUserForm from '../profile-user-form/profile-user-form';
+import ProfileOrders from '../profile-order/profile-order';
+import IngredientDetails from '../ingredient-details/ingredient-details';
+import Modal from '../modal/modal';
 import {
   TO_LOGIN,
   TO_FORGOT_PASSWORD,
@@ -23,18 +23,18 @@ import {
   TO_RESET_PASSWORD,
   TO_INGREDIENTS,
   TO_ORDERS,
-} from "../../utils/route-constants";
-import { getUser } from "../../services/auth/auth-actions";
-import { getCookie } from "../../utils/burger-api";
+} from '../../utils/route-constants';
+import { getUser } from '../../services/auth/auth-actions';
+import { getCookie } from '../../utils/burger-api';
 
-const App = ():JSX.Element => {
+const App = (): JSX.Element => {
   const dispatch = useDispatch() as any;
   const location = useLocation();
   const background = location.state && location.state.background;
 
   useEffect(() => {
-    const accessToken = getCookie("accessToken");
-    const refreshToken = sessionStorage.getItem("refreshToken");
+    const accessToken = getCookie('accessToken');
+    const refreshToken = sessionStorage.getItem('refreshToken');
     if (accessToken && refreshToken) {
       dispatch(getUser());
     }
@@ -42,7 +42,7 @@ const App = ():JSX.Element => {
   return (
     <>
       <Routes location={background || location}>
-        <Route path="/" element={<Layout />}>
+        <Route path='/' element={<Layout />}>
           <Route index element={<BurgerPage />} />
           <Route path={TO_LOGIN} element={<LoginPage />} />
           <Route path={TO_REGISTR} element={<RegisterPage />} />
@@ -54,15 +54,14 @@ const App = ():JSX.Element => {
           />
           <Route
             path={TO_PROFILE}
-            element={<ProtectedRoute element={<ProfilePage />} />}
-          >
+            element={<ProtectedRoute element={<ProfilePage />} />}>
             <Route index element={<ProfileUserForm />} />
             <Route
               path={`${TO_PROFILE}${TO_ORDERS}`}
               element={<ProfileOrders />}
             />
           </Route>
-          <Route path="*" element={<NotFoundPage />} />
+          <Route path='*' element={<NotFoundPage />} />
         </Route>
       </Routes>
       {background && (
