+++ src/components/error-bage/error-bage.tsx
@@ -1,15 +1,15 @@
-import PropTypes from 'prop-types';
 import styles from './error-bage.module.css';
-const ErrorBage = ({ error }) => {
+
+type TProps = {
+  error: string;
+};
+const ErrorBage = ({ error }: TProps): JSX.Element => {
   return (
     <div className={styles.container}>
       <p className='text text_type_main-large mb-4'>{error}</p>
       <p className='text text_type_main-medium'>Kажется что-то пошло не так!</p>
     </div>
   );
-};
-ErrorBage.propTypes = {
-  error: PropTypes.string.isRequired,
 };
 
 export default ErrorBage;
