+++ src/components/ingredient-details/ingredient-details.tsx
@@ -1,18 +1,18 @@
-import { useSelector } from "react-redux";
-import styles from "./ingredient-details.module.css";
-import { getUiState } from "../../services/ui/ui-selectors";
+import { useSelector } from 'react-redux';
+import styles from './ingredient-details.module.css';
+import { getUiState } from '../../services/ui/ui-selectors';
 
 const NUTRIENTS = [
   {
-    name: "calories",
-    text: "Калории, ккал",
+    name: 'calories',
+    text: 'Калории, ккал',
   },
-  { name: "proteins", text: "Белки, г" },
-  { name: "fat", text: "Жиры, г" },
-  { name: "carbohydrates", text: "Углеводы, г" },
+  { name: 'proteins', text: 'Белки, г' },
+  { name: 'fat', text: 'Жиры, г' },
+  { name: 'carbohydrates', text: 'Углеводы, г' },
 ];
 
-const IngredientDetails = () => {
+const IngredientDetails = (): JSX.Element => {
   const { ingredient } = useSelector(getUiState);
   return (
     ingredient && (
@@ -21,7 +21,7 @@ const IngredientDetails = () => {
           <img
             className={styles.image}
             src={ingredient.image_large}
-            alt="ingredient"
+            alt='ingredient'
           />
         </div>
         <div>
@@ -33,15 +33,14 @@ const IngredientDetails = () => {
           {NUTRIENTS.map((element, index) => (
             <div
               className={`${styles.infoContaienr} pt-5`}
-              key={index + element.text}
-            >
+              key={index + element.text}>
               <div>
-                <p className="text text_type_main-default text_color_inactive">
+                <p className='text text_type_main-default text_color_inactive'>
                   {element.text}
                 </p>
               </div>
               <div>
-                <p className="text text_type_digits-default text_color_inactive">
+                <p className='text text_type_digits-default text_color_inactive'>
                   {ingredient[element.name]}
                 </p>
               </div>
