+++ src/components/layout/layout.tsx
@@ -2,7 +2,7 @@ import { Outlet } from 'react-router-dom';
 import AppHeader from '../app-header/app-header';
 import styles from './layout.module.css';
 
-const Layout = () => {
+const Layout = (): JSX.Element => {
   return (
     <div className={styles.layout}>
       <AppHeader />
