+++ src/components/loading-bage/loading-bage.tsx
@@ -1,5 +1,6 @@
 import styles from './loading-bage.module.css';
-const LoadingBage = () => {
+
+const LoadingBage = (): JSX.Element => {
   return (
     <div className={styles.container}>
       <p className='text text_type_main-large'>Loading...</p>
