--- src/components/modal/model-overlay/model-overlay.js
@@ -1,12 +1,0 @@
-import PropTypes from 'prop-types';
-import styles from './model-overlay.module.css';
-
-const ModelOverlay = ({ onClose }) => (
-  <div className={styles.modalOverlay} onClick={onClose} />
-);
-
-ModelOverlay.propTypes = {
-  onClose: PropTypes.func.isRequired,
-};
-
-export default ModelOverlay;
