+++ src/components/profile-order/profile-order.tsx
@@ -1,4 +1,4 @@
-const ProfileOrders = () => {
+const ProfileOrders = (): JSX.Element => {
   return (
     <div>
       <p className='text text_type_main-large'>PROFILE ORDERS</p>
